+++ packages/vite-plugin-inspect-react/CHANGELOG.md
@@ -4,13 +4,16 @@
 
 ### Minor Changes
 
-- [`f27d25d`](https://github.com/willnguyen1312/namnode/commit/f27d25d42aecf04477ef69edd6d5caccc3551d62) Thanks [@willnguyen1312](https://github.com/willnguyen1312)! - Add documentation to READ and improve inspector's result 💞
+- [`f27d25d`](https://github.com/willnguyen1312/namnode/commit/f27d25d42aecf04477ef69edd6d5caccc3551d62) Thanks
+  [@willnguyen1312](https://github.com/willnguyen1312)! - Add documentation to READ and improve inspector's result 💞
 
 ## 1.1.0
 
 ### Minor Changes
 
-- [#13](https://github.com/willnguyen1312/namnode/pull/13) [`7acb4a8`](https://github.com/willnguyen1312/namnode/commit/7acb4a816b7c26cab40c5f100d122c1d1f764a70) Thanks [@willnguyen1312](https://github.com/willnguyen1312)! - Enhance vite plugin inspect react 💞
+- [#13](https://github.com/willnguyen1312/namnode/pull/13)
+  [`7acb4a8`](https://github.com/willnguyen1312/namnode/commit/7acb4a816b7c26cab40c5f100d122c1d1f764a70) Thanks
+  [@willnguyen1312](https://github.com/willnguyen1312)! - Enhance vite plugin inspect react 💞
 
 ## 1.0.0